requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
coloredlogs==15.0.1
ConfigArgParse==1.5.3
cvu-python==0.0.2
dlib==19.15.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.89.1
filelock==3.9.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==23.1.21
gdown==4.6.3
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
locust==2.14.2
MarkupSafe==2.1.2
mpmath==1.2.1
msgpack==1.0.4
numpy==1.24.1
onnxruntime==1.14.0
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==4.22.0
psutil==5.9.4
pydantic==1.10.4
PySocks==1.7.1
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
sympy==1.11.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
vidsz==0.2.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.14.0
zope.event==4.6
zope.interface==5.5.2